requirements.txt
python-dotenv
pandas
numpy
scikit-learn
matplotlib
seaborn
scipy
Flask
pymongo
certifi
pymongo[srv]>=4.3.3
pymongo
dill
pyaml
mlflow
dagshub
fastapi
uvicorn
python-multipart
setuptools